test: updated how window object is assigned store
